refactor(cli): extract doctor handler to src/cli/commands/doctor.mjs
main() の doctor 分岐を runDoctorCommand(parsed, targetPath) へ抽出（#1592 と同型）。
dispatch は return await で main() の outer try/catch を経由させ、非 git ディレクトリでの
friendly エラー（Not a git repository + Hints, exit 1）を維持する。
doctorLocalReview / isLlmEnabled / MAX_DIFF_PREVIEW_LINES の import をハンドラ側へ移す。
振る舞い・出力・exit code は不変。
